backend/requirements.txt
fastapi
uvicorn[standard]
pillow
requests
pandas
openpyxl
numpy
